[JIT] Make type singletons const references
**Summary**
Certain types that cannot be specialized (e.g. `IntType`) have
singletons that are used to represent them in JIT in order to avoid
making and maintaining multiple copies of what is semantically the same
object. These singletons are accessible through `XXXType::get()`.
However, these functions return `XXXTypePtr`, which is an alias for
`std::shared_ptr<XXXType>`, and can therefore be modified by the caller.
This is dangerous, so this commit makes these `get()` functions return
`const XXXTypePtr&` so that they cannot be modified, and so that a new
`std::shared_ptr` is not constructed to return the singleton.

**Test Plan**
Existing unit tests.

ghstack-source-id: d5aba19457891d35d50cedc6e3bf983a0d28a22d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48027