T78750158 Support varying size input in numeric suite at 10/30/2020, 3:55:01 PM (#47391)
* T78750158 Support varying size input in numeric suite at 10/30/2020, 3:55:01 PM

Current Numeric Suite will fail if it's collecting for multiple inputs and each input is of not same size.  This fix adds support for varying size input in numeric suite.

Differential Revision: [D24662271](https://our.internmc.facebook.com/intern/diff/D24662271/)

[ghstack-poisoned]

* Update on "T78750158 Support varying size input in numeric suite at 10/30/2020, 3:55:01 PM"


Current Numeric Suite will fail if it's collecting for multiple inputs and each input is of not same size.  This fix adds support for varying size input in numeric suite.

Differential Revision: [D24662271](https://our.internmc.facebook.com/intern/diff/D24662271/)

[ghstack-poisoned]

Co-authored-by: annieguan <annieguan@devvm2014.vll0.facebook.com>