Don't store device indices separately on FutureNCCL
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48501

This commit is part of a stack that reworks FutureNCCL in order to extract a generic CUDA-aware Future subclass. The stack deliberately breaks up this transition into elementary changes, to make it easier to verify that the behavior is preserved (or to highlight how it gets changed).

---

FutureNCCL stores a set of devices (on which the tensors in the data reside) and a CUDA event for each of those devices. In fact, each event instance also already contains the device it belongs to, which means we can avoid storing that information separately (with the risk that it'll be mismatched and/or inaccurate).
ghstack-source-id: 117420190

Differential Revision: [D25177554](https://our.internmc.facebook.com/intern/diff/D25177554/)