[PyTorch] Avoid refcount decrement in alias_with_sizes_and_strides
The previous code created a temporary `Tensor`, which means
that we had an atomic reference count decrement (expensive!) to
destroy its inner `c10::intrusive_ptr`. Now, we instead create a
temporary `c10::intrusive_ptr` that we ensure is null by destruction
time, avoiding the expensive atomic operation. (`self_` should not be
destroyed because of NRVO.)

Differential Revision: [D25258631](https://our.internmc.facebook.com/intern/diff/D25258631/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D25258631/)!

ghstack-source-id: 117608962
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48678