VariableType calls faithful C++ API for c10-full out ops
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47792

For operators with out arguments, VariableType previously called the out overload of the C++ API because that's all we had.
We introduced a faithful C++ API that takes out arguments in schema-order in D24835252 and this PR changes VariableType to use that API instead.

Note that this only applies to c10-full ops. Non-c10-full ops still call the unfaithful API. There aren't any c10-full out ops at the moment.
So this PR can only be tested and evaluated together with PRs on top that make ops with out arguments c10-full.
ghstack-source-id: 118068088

Differential Revision: [D24901945](https://our.internmc.facebook.com/intern/diff/D24901945/)