[JIT] Use `is_buffer` in `BufferPolicy::valid`
**Summary**
`BufferPolicy::valid` uses `!typ->is_parameter(i)` to check if an
attribute is a buffer or not; it should use `type->is_buffer(i)` instead.

**Test Plan**
It is difficult to write an additional test that would have failed before this
commit because the two booleans `is_parameter` and `is_buffer` are never set
to `true` at the same time.

**Fixes**
This commit fixes #48746.

ghstack-source-id: e249dd1a866b6bcb2fb58488b361601bcac48309
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49053