[PyTorch] Remove aten::native::empty usage in TensorIndexing
Try to resolve part of the github issue of https://github.com/pytorch/pytorch/issues/48684 . ```aten::native::empty()``` is referenced in TensorIndexing.h. However, the definition of ```aten::native::empty()``` is nothing but checks and eventually calling ```at::empty()```.

In this diff, ```at::empty()``` is directly used to avoid the reference to native symbols.

Differential Revision: [D25417854](https://our.internmc.facebook.com/intern/diff/D25417854/)

ghstack-source-id: 118165999
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49074